+++ util/jvmlib/src/commonMain/kotlin/io/github/shakelang/jvmlib/infos/constants/ConstantFloatInfo.kt
@@ -11,7 +11,7 @@ class ConstantFloatInfo(val value: Float) : ConstantInfo() {
     override val tag: Byte get() = ConstantFloatInfo.tag
     override val tagName: String get() = name
 
-    override fun toJson() = super.toJson().with("value", value)
+    override fun toJson() = super.toJson().with("value", value.toDouble())
 
     override fun dump(out: DataOutputStream) {
         out.writeByte(tag)
